move BTAnalyticsEvent and BTAnalyticsSession structs into BTAnalyticsService
