--- Sources/BraintreeCoreSwift/BTAnalyticsEvent.swift
@@ -1,20 +1,0 @@
-import Foundation
-
-/// Encapsulates a single analytics event
-struct BTAnalyticsEvent {
-
-    var eventName: String
-    var timestamp: UInt64
-
-    var description: String {
-        "\(eventName) at \(timestamp)"
-    }
-
-    /// Event serialized to JSON
-    var json: [String: Any] {
-        [
-            "eventName": eventName,
-            "timestamp": timestamp
-        ]
-    }
-}
