--- Sources/BraintreeCoreSwift/BTAnalyticsSession.swift
@@ -1,14 +1,0 @@
-import Foundation
-
-/// Encapsulates analytics events for a given session
-struct BTAnalyticsSession {
-
-    let sessionID: String
-    let source: String
-    let integration: String
-
-    var events: [BTAnalyticsEvent] = []
-
-    /// Dictionary of analytics metadata from `BTAnalyticsMetadata`
-    let metadataParameters: [String: Any] = BTAnalyticsMetadata.metadata
-}
